Move the Tycho Version to the SPI
