+++ tycho-core/src/test/java/org/eclipse/tycho/core/test/DependencyComputerTest.java
@@ -48,8 +48,8 @@ import org.eclipse.tycho.core.osgitools.DependencyComputer.DependencyEntry;
 import org.eclipse.tycho.core.osgitools.EquinoxResolver;
 import org.eclipse.tycho.core.utils.MavenSessionUtils;
 import org.eclipse.tycho.core.utils.TychoProjectUtils;
-import org.eclipse.tycho.core.utils.TychoVersion;
 import org.eclipse.tycho.testing.AbstractTychoMojoTestCase;
+import org.eclipse.tycho.version.TychoVersion;
 import org.junit.Assert;
 import org.junit.Test;
 import org.osgi.framework.BundleException;
