+++ tycho-core/src/test/java/org/eclipse/tycho/core/utils/TychoVersionTest.java
@@ -19,6 +19,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
 
+import org.eclipse.tycho.version.TychoVersion;
 import org.junit.Test;
 
 public class TychoVersionTest {
