+++ tycho-spi/src/main/resources/org/eclipse/tycho/version/version.properties
